Handling ZWave CentralScene command class events
